requirements.txt
gradio==5.0.1
langchain==0.1.11
langchain-anthropic==0.1.4
anthropic==0.18.1
python-dotenv==1.0.1
pypdf==4.0.1